requirements.txt
web3==6.0.0
eth-account==0.8.0
requests==2.28.2
rich==13.3.4
python-dotenv==1.0.0
 